split output

+++ messageReceiver_processes.py
@@ -6,8 +6,8 @@ Multiprocessing also spawns new processes, but they run your code, and are desig
 import shlex
 import boto3
 import time
-
-import config #config.py confgiuration file
+import json
+import config #config.py 
 
 env             = 'DEV'
 max_q_messages  = config.SQS_CONFIG[env]['max_messages_received']
@@ -30,23 +30,25 @@ session         = boto3.Session(**session_cfg)
 sqs             = session.resource('sqs',**sqs_cfg)
 queue           = sqs.get_queue_by_name(QueueName=queue_name)
 
-# Get messages until finished
+## Get messages until finished
 while True:
-    messages  = queue.receive_messages(MaxNumberOfMessages = max_q_messages)
+    messages = queue.receive_messages(MaxNumberOfMessages=max_q_messages)
     for message in messages:
-	    # Process messages by printing out body
-		print('Hello, {0}'.format(message.body))
+		print('{0}, {1}, {2}'.format(message.body, message.message_id, message.receipt_handle))
  		args = shlex.split(message.body)
  		if args[0].endswith(".php"): 
 			cmd = "php " + message.body
 			process = Popen(cmd, shell=True, stdout=PIPE, stderr=PIPE)
 			stdout, stderr = process.communicate()
-			print("stdout = " + stdout)
-			print("stderr = " + stderr)
-
-			if (sterrd):
-				# Move to dead letter queue, with the stdError data as metadata!
-				
-
-		# Let the queue know that the message is processed
-		message.delete()
+			
+			if (stderr):
+				## Move to dead letter queue, with the stdError data as metadata!
+				"""
+				The message attributes of an SQS message are immutable once the message has been sent to the queue. The SQS Query API (used by all client libraries) has no support for modifying a message in the queue, other than to change its visibility timeout.
+				"""
+				receipt = message.receipt_handle
+				print('Processing error, {id}, {body}, with error: {error}'.format(body=message.body, id=message.message_id, error= stderr))
+			else:	
+				print("stdout = " + stdout)	
+				## Let the queue know that the message is processed
+				message.delete()
